# docs/swagger.yaml
basePath: /
definitions:
  doc.CPDetail:
    properties:
      keterangan:
        type: string
      name:
        type: string
      ok:
        type: boolean
    type: object
  doc.DeviceDocForm:
    properties:
      checkpoint:
        items:
          $ref: '#/definitions/doc.CPDetail'
        type: array
      device_id:
        type: string
    type: object
  doc.UpdateDeviceDocForm:
    properties:
      checkpoint:
        items:
          $ref: '#/definitions/doc.CPDetail'
        type: array
    type: object
  handler.cctvForm:
    properties:
      kode:
        type: string
      lokasi:
        type: string
      nama:
        type: string
    type: object
  handler.fingerPrintForm:
    properties:
      kode:
        type: string
      lokasi:
        type: string
      nama:
        type: string
    type: object
...
  handler.updateCheckpointForm:
    properties:
      checkpoint:
        items:
          type: string
        type: array
    type: object
...
info:
  contact: {}
  description: Sistem Pencatatan Maintenance IT Backend API
  title: Sistem Pencatatan Maintenance IT Backend
paths:
  /api/cctv:
    post:
      operationId: create-cctv
      parameters:
      - description: CCTV Form
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.cctvForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Create cctv
      tags:
      - Device CCTV
  /api/cctv/{id}:
    delete:
      operationId: delete-cctv-by-id
      parameters:
      - description: CCTV ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Delete cctv by id
      tags:
      - Device CCTV
    get:
      operationId: get-cctv-by-id
      parameters:
      - description: CCTV ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Get cctv by id
      tags:
      - Device CCTV
    put:
      operationId: update-cctv-by-id
      parameters:
      - description: CCTV ID
        in: path
        name: id
        required: true
        type: string
      - description: CCTV Form
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.cctvForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Update cctv by id
      tags:
      - Device CCTV
  /api/cctvs:
    get:
      operationId: get-all-cctvs
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Get all cctvs
      tags:
      - Device CCTV
  /api/checkpoint/cctv:
    get:
      operationId: get-cctv-checkpoint
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Get cctv checkpoint
      tags:
      - Checkpoint
    put:
      consumes:
      - application/json
      operationId: update-cctv-checkpoint
      parameters:
      - description: Checkpoint
        in: body
        name: checkpoint
        required: true
        schema:
          $ref: '#/definitions/handler.updateCheckpointForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Update cctv checkpoint
      tags:
      - Checkpoint
  /api/checkpoint/fingerprint:
    get:
      operationId: get-fingerprint-checkpoint
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Get fingerprint checkpoint
      tags:
      - Checkpoint
    put:
      consumes:
      - application/json
      operationId: update-fingerprint-checkpoint
      parameters:
      - description: Checkpoint
        in: body
        name: checkpoint
        required: true
        schema:
          $ref: '#/definitions/handler.updateCheckpointForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Update fingerprint checkpoint
      tags:
      - Checkpoint
...
  /api/doc/cctv:
    post:
      operationId: create-cctv-document
      parameters:
      - description: CCTV Document Form
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/doc.DeviceDocForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Create cctv document
      tags:
      - Doc CCTV
  /api/doc/cctv/{id}:
    delete:
      operationId: delete-cctv-document-by-id
      parameters:
      - description: CCTV Document ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Delete cctv document by ID
      tags:
      - Doc CCTV
    get:
      operationId: get-cctv-document-by-id
      parameters:
      - description: CCTV Document ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Get cctv document by ID
      tags:
      - Doc CCTV
    put:
      operationId: update-cctv-document-by-id
      parameters:
      - description: CCTV Document ID
        in: path
        name: id
        required: true
        type: string
      - description: CCTV Document Form
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/doc.UpdateDeviceDocForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Update cctv document by ID
      tags:
      - Doc CCTV
  /api/doc/cctvs:
    get:
      operationId: get-all-cctv-documents
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Get all cctv documents
      tags:
      - Doc CCTV
  /api/doc/fingerprint:
    post:
      operationId: create-fingerprint-document
      parameters:
      - description: Fingerprint Document Form
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/doc.DeviceDocForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Create fingerprint document
      tags:
      - Doc Fingerprint
  /api/doc/fingerprint/{id}:
    delete:
      operationId: delete-fingerprint-document-by-id
      parameters:
      - description: Fingerprint Document ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Delete fingerprint document by ID
      tags:
      - Doc Fingerprint
    get:
      operationId: get-fingerprint-document-by-id
      parameters:
      - description: Fingerprint Document ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Get fingerprint document by ID
      tags:
      - Doc Fingerprint
    put:
      operationId: update-fingerprint-document-by-id
      parameters:
      - description: Fingerprint Document ID
        in: path
        name: id
        required: true
        type: string
      - description: Fingerprint Document Form
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/doc.UpdateDeviceDocForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Update fingerprint document by ID
      tags:
      - Doc Fingerprint
  /api/doc/fingerprints:
    get:
      operationId: get-all-fingerprint-documents
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Get all fingerprint documents
      tags:
      - Doc Fingerprint
...
  /api/fingerprint:
    post:
      operationId: create-fingerprint
      parameters:
      - description: Fingerprint Form
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.fingerPrintForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Create new fingerprint
      tags:
      - Device Fingerprint
  /api/fingerprint/{id}:
    delete:
      operationId: delete-fingerprint-by-id
      parameters:
      - description: Fingerprint ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Delete fingerprint by id
      tags:
      - Device Fingerprint
    get:
      operationId: get-fingerprint-by-id
      parameters:
      - description: Fingerprint ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Get fingerprint by id
      tags:
      - Device Fingerprint
    put:
      operationId: update-fingerprint-by-id
      parameters:
      - description: Fingerprint ID
        in: path
        name: id
        required: true
        type: string
      - description: Fingerprint Form
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.fingerPrintForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Update fingerprint by id
      tags:
      - Device Fingerprint
  /api/fingerprints:
    get:
      operationId: get-all-fingerprints
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Get all fingerprints
      tags:
      - Device Fingerprint
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"